Add: Store HSB and HSL hues in range 0 - 360
After reading up on the colour formats it seems the hues are typically stored in the range 0 - 360 but previously the code allowed this range as well as 0 - 1. The other values are typically stored in the range 0 - 1 but previously the code allowed this range as well as 0 - 100. The trouble being that these ranges can overlap. This overlap only occurs at extremes but still occurs. For example HSB(0, 1, 1) could be an unshaded and untinted red if all values are in the 0 - 1 range. But it could also be almost black if the hue is in the range 0 - 360 and saturation and brightness in the range 0 - 100. For this reason I've gone with the typical values only.