Add: Move getRGB and toHex into Colour class
The getRGB method is now represented by the Colour class' constructor. The difference though is that hsb and hsl representations get returned as such instead of being converted to RGB as they were previously. This may require adding toRGB to the RGB class, toHSB to the HSB class and toHSL to the HSL class in order that an ambiguous return value from the Colour class can be made unambiguous.

Some corrections were also made to the RegEx which was seen to not be working correctly when testing (only) in Firefox.