Add: Conversion methods to Colour class
When a colour is instantiated from a string we don't know its type. This means we need methods to convert to HSL, HSB and RGB. We want all three colour classes to inherit from Colour and so we need to ensure the three conversion methods behave as expected for all three classes.